Machine search/requirements.txt
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.131.0
Pillow==10.3.0
imagehash==4.3.1
pdfplumber==0.11.0
PyPDF2==3.0.1
rapidfuzz==3.9.3
python-magic-bin==0.4.14
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
opencv-python==4.9.0.80
sentence-transformers==3.0.1
torch==2.3.1
transformers==4.41.2
scikit-learn==1.5.0
chardet==5.2.0
python-docx==1.1.2
striprtf==0.0.26
openpyxl==3.1.4
tqdm==4.66.4
